refactor(frontend): rename useClosePane to plain closePane function
